Update merge strategy, split and extend merging tests
