+++ ariadne_graphql_proxy/__init__.py
@@ -1,5 +1,37 @@
 from .copy import copy_schema
+from .merge import (
+    merge_args,
+    merge_enums,
+    merge_enums_values,
+    merge_fields,
+    merge_inputs,
+    merge_input_fields,
+    merge_interfaces,
+    merge_objects,
+    merge_scalars,
+    merge_schemas,
+    merge_type_maps,
+    merge_types,
+    merge_unions,
+)
 from .proxy_resolver import setup_root_resolver
 from .remote_schema import GraphQLRemoteSchema
 
-__all__ = ["GraphQLRemoteSchema", "copy_schema", "setup_root_resolver"]
+__all__ = [
+    "GraphQLRemoteSchema",
+    "copy_schema",
+    "merge_args",
+    "merge_enums",
+    "merge_enums_values",
+    "merge_fields",
+    "merge_inputs",
+    "merge_input_fields",
+    "merge_interfaces",
+    "merge_objects",
+    "merge_scalars",
+    "merge_schemas",
+    "merge_type_maps",
+    "merge_types",
+    "merge_unions",
+    "setup_root_resolver",
+]
